BUILD: user verwijderen functie toegevoegd

+++ src/main/java/com/ipsen/spine/model/User.java
@@ -21,7 +21,7 @@ import java.util.List;
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
     private String firstName;
     private String lastName;
     @Column(unique = true)
@@ -33,6 +33,7 @@ public class User implements UserDetails {
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return List.of(new SimpleGrantedAuthority(role.name()));
     }
+
 
     @Override
     public String getUsername() {
